% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pressuRe_functions.R
\name{arch_index}
\alias{arch_index}
\title{Calculate Arch Index.}
\usage{
arch_index(pressure_data, plot = TRUE)
}
\arguments{
\item{pressure_data}{List. Includes a 3D array covering each timepoint of the
measurement. z dimension represents time}

\item{plot}{Logical. Not implemented yet}
}
\value{
Numeric. Arch index value
}
\description{
Calculate Arch Index.
}
\examples{
emed_data <- system.file("extdata", "emed_test.lst", package = "pressuRe")
pressure_data <- load_emed(emed_data)
arch_index(pressure_data)
}
