% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pressuRe_functions.R
\name{area_curve}
\alias{area_curve}
\title{Contact area curve}
\usage{
area_curve(pressure_data, threshold = 0, plot = FALSE)
}
\arguments{
\item{pressure_data}{List. First item should be a 3D array covering each
timepoint of the measurement. z dimension represents time}

\item{threshold}{Numeric. The minimum pressure required for a sensor to be
considered active}

\item{plot}{Logical. If TRUE also plots data as line curve}
}
\value{
Numeric vector containing force values
}
\description{
Generate contact area curve with option to plot
}
\examples{
emed_data <- system.file("extdata", "emed_test.lst", package = "pressuRe")
pressure_data <- load_emed(emed_data)
area_curve(pressure_data, plot = TRUE)
}
\author{
Scott Telfer \email{scott.telfer@gmail.com}
}
