% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pressuRe_functions.R
\name{clean_pressure}
\alias{clean_pressure}
\title{Manually remove individual sensors}
\usage{
clean_pressure(pressure_data)
}
\arguments{
\item{pressure_data}{List. First item should be a 2D array covering each
timepoint of the measurement. z dimension represents time.}
}
\value{
\itemize{
  \item pressure_array. 2D array covering each timepoint of the measurement.
           Sensors are columns, time is rows
  \item pressure_system. String defining pressure system
  \item sens_size. Numeric vector with the dimensions of the sensors
  \item time. Numeric value for time between measurements
  \item masks. List
  \item events. List
  \item sensor_polygons. Data frame with corners of sensors
  \item max_matrix. Matrix
 }
}
\description{
Select sensors, often with abherrant readings, and remove from
the dataset
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
emed_data <- system.file("extdata", "emed_test.lst", package = "pressuRe")
pressure_data <- load_emed(emed_data)
pressure_data <- clean_pressure(pressure_data)
\dontshow{\}) # examplesIf}
}
