% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pressuRe_functions.R
\name{cop}
\alias{cop}
\title{Center of pressure}
\usage{
cop(pressure_data)
}
\arguments{
\item{pressure_data}{List. First item is a 3D array covering each timepoint
of the measurement. z dimension represents time}
}
\value{
Data frame with x and y coordinates of COP throughout trial
}
\description{
Generates center of pressure coordinates
}
\examples{
emed_data <- system.file("extdata", "emed_test.lst", package = "pressuRe")
pressure_data <- load_emed(emed_data)
cop(pressure_data)
}
\author{
Scott Telfer \email{scott.telfer@gmail.com}
}
