% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pressuRe_functions.R
\name{create_mask}
\alias{create_mask}
\title{Create mask}
\usage{
create_mask(pressure_data, n_verts = 4, image = "max", preview = TRUE)
}
\arguments{
\item{pressure_data}{List. First item is a 3D array covering each timepoint
of the measurement.}

\item{n_verts}{Numeric. Number of vertices in mask}

\item{image}{String. "max" = footprint of maximum sensors. "mean"
average value of sensors over time (usually for static analyses)}

\item{preview}{Logical. Show new maks on pressure image}
}
\value{
List New mask is added to the relevant A 3D array covering each timepoint of the measurement for the
  selected region. z dimension represents time
}
\description{
Manually create mask region
}
\examples{
emed_data <- system.file("extdata", "emed_test.lst", package = "pressuRe")
pressure_data <- load_emed(emed_data)
pressure_data <- create_mask(pressure_data, 4)
}
\author{
Scott Telfer \email{scott.telfer@gmail.com}
}
