% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pressuRe_functions.R
\name{force_curve}
\alias{force_curve}
\title{Force curve}
\usage{
force_curve(pressure_data, plot = FALSE)
}
\arguments{
\item{plot}{Logical. If TRUE also plots data as line curve}

\item{pressure_data.}{List. A 3D array covering each timepoint of the
measurement. z dimension represents time}
}
\value{
Numeric vector containing force values
}
\description{
Generates force vector with option to plot
}
\examples{
force_curve(pressure_data, plot = TRUE)
}
\author{
Scott Telfer \email{scott.telfer@gmail.com}
}
