% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pressuRe_functions.R
\name{force_pedar}
\alias{force_pedar}
\title{pedar force}
\usage{
force_pedar(pressure_data, foot_side)
}
\arguments{
\item{pressure_data}{List.}

\item{foot_side}{String}
}
\description{
pedar force
}
