% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pressuRe_functions.R
\name{getMinBBox}
\alias{getMinBBox}
\title{Get minimum bounding box (angle)}
\usage{
getMinBBox(xy)
}
\arguments{
\item{xy}{Matrix. n row by 2 column with x-y coords of points to be bounded}
}
\description{
Get minimum bounding box (angle)
}
