% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pressuRe_functions.R
\name{gglocator}
\alias{gglocator}
\title{gglocator}
\usage{
gglocator(
  n = 1,
  message = FALSE,
  xexpand = c(0, 0),
  yexpand = c(0, 0),
  mercator = TRUE
)
}
\arguments{
\item{n}{Integer. Number of points to select}
}
\value{
Data frame. x and y coordinates of selected point
}
\description{
gglocator
}
