% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pressuRe_functions.R
\name{load_fscan}
\alias{load_fscan}
\title{Load F-scan data}
\usage{
load_fscan(pressure_filepath)
}
\arguments{
\item{pressure_filepath}{String. Filepath pointing to emed pressure file}
}
\value{
A list with information about the pressure data.
\itemize{
  \item pressure_array. A 3D array covering each timepoint of the measurement.
           z dimension represents time
  \item sens_size. Numeric vector with two values for the dimensions of the sensors
  \item time. Numeric value for time between measurements
 }
}
\description{
Imports and formats .asc files collected on pedar system and
   exported from Tekscan software
}
\author{
Scott Telfer \email{scott.telfer@gmail.com}
}
