% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pressuRe_functions.R
\name{load_iscan}
\alias{load_iscan}
\title{Load tekscan i-scan data}
\usage{
load_iscan(pressure_filepath, sensor_type, sensor_pad)
}
\arguments{
\item{pressure_filepath}{String. Filepath pointing to emed pressure file}

\item{sensor_type}{String. Currently only "6900" supported}

\item{sensor_pad}{Integer. If sensor has multiple sensor pads choose the one
of interest}
}
\value{
Array. A 3D array covering each timepoint of the measurement. z
  dimension represents time
}
\description{
Load tekscan i-scan data
}
\author{
Scott Telfer \email{scott.telfer@gmail.com}
}
