% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pressuRe_functions.R
\name{load_stappone}
\alias{load_stappone}
\title{Load stappone data}
\usage{
load_stappone(pressure_filepath)
}
\arguments{
\item{pressure_filepath}{String. Filepath pointing to stappone pressure file}
}
\value{
A list with information about the pressure data.
\itemize{
  \item pressure_array. 2D array covering each timepoint of the measurement.
           Sensors are columns, time is rows
  \item pressure_system. String defining pressure system
  \item sens_size. String with sensor type
  \item time. Numeric value for time between measurements
  \item masks. List
  \item events. List
  \item sensor_polygons. Data frame with corners of sensors
  \item max_matrix Matrix with maximum image
 }
}
\description{
Imports and formats .asc files collected on stappone system
}
\examples{
stappone_data <- system.file("extdata", "stappone_example.csv", package = "pressuRe")
pressure_data <- load_stappone(stappone_data)
}
