% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pressuRe_functions.R
\name{load_tekscan}
\alias{load_tekscan}
\title{Load F-scan data}
\usage{
load_tekscan(pressure_filepath)
}
\arguments{
\item{pressure_filepath}{String. Filepath pointing to emed pressure file}
}
\value{
A list with information about the pressure data.
\itemize{
  \item pressure_array. 3D array covering each timepoint of the measurement.
           z dimension represents time
  \item pressure_system. String defining pressure system
  \item sens_size. Numeric vector with the dimensions of the sensors
  \item time. Numeric value for time between measurements
  \item masks. List
  \item events. List
 }
 @examples
tekscan_data <- system.file("extdata", "fscan_testL.asf", package = "pressuRe")
tekscan_data <- system.file("extdata", "iscan_test.csv", package = "pressuRe")
pressure_data <- load_tekscan(tekscan_data)
 @importFrom
 @export
}
\description{
Imports and formats .asc files collected on f-scan system and
   exported from Tekscan software
}
\author{
Scott Telfer \email{scott.telfer@gmail.com}
}
