% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pressuRe_functions.R
\name{mask_analysis}
\alias{mask_analysis}
\title{Analyze masked regions of pressure data}
\usage{
mask_analysis(
  pressure_data,
  masks,
  partial_sensors = FALSE,
  variable = "peak_sensor"
)
}
\arguments{
\item{pressure_data}{List. Includes a 3D array covering each timepoint of the
measurement. z dimension represents time}

\item{masks}{List. Masks used to define the regions to be analysed}

\item{partial_sensors}{Logical Defines how sensors that do not
lie wholly within mask are dealt with. If FALSE, they will be excluded;
if TRUE, for relevant variables their contribution will be weighted by the
proportion of the sensor that falls within the mask border}

\item{variable}{String. Variable to be determined. "peak_sensor",
"peak_sensor_ts", "peak_mask", "peak_mask_ts", "mean_mask",
"contact_area_peak", "contact_area_ts", "pti_1", "pti_2", force_peak",
"force_ts"}
}
\value{
Data frame.
}
\description{
Analyze masked regions of pressure data
}
\author{
Scott Telfer \email{scott.telfer@gmail.com}
}
