% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pedar_insole_areas}
\alias{pedar_insole_areas}
\title{pedar sensor size data}
\format{
## `pedar_insole_areas`
A data frame with 198 rows and 8 columns:
\describe{
  \item{u}{areas for size u}
  \item{v}{areas for size v}
  \item{w}{areas for size w}
  \item{x}{areas for size x}
  \item{y}{areas for size y}
  \item{uw}{areas for size uw}
  \item{xw}{areas for size xw}
  \item{vw}{areas for size vw}
}
}
\source{
Scott Telfer
}
\usage{
pedar_insole_areas
}
\description{
Sensor sizes for different pedar insoles
}
\keyword{datasets}
