% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pedar_insole_grid}
\alias{pedar_insole_grid}
\title{pedar sensor grid data}
\format{
## `pedar_insole_grid`
A data frame with 199 rows and 8 columns:
\describe{
  \item{V1}{x1 coordinate}
  \item{V2}{y1 coordinate}
  \item{V3}{x2 coordinate}
  \item{V4}{y2 coordinate}
  \item{V5}{x3 coordinate}
  \item{V6}{y3 coordinate}
  \item{V7}{x4 coordinate}
  \item{V8}{y4 coordinate}
}
}
\source{
Scott Telfer
}
\usage{
pedar_insole_grid
}
\description{
Sensor outline coordinates
}
\keyword{datasets}
