% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pressuRe_functions.R
\name{plot_pedar}
\alias{plot_pedar}
\title{plot pedar}
\usage{
plot_pedar(pressure_data, pressure_image = "step_max", step_n)
}
\arguments{
\item{pressure_data}{List.}

\item{pressure_image}{String.}

\item{step_n}{Numeric}
}
\description{
plot pedar
}
