% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pressuRe_functions.R
\name{plot_pressure}
\alias{plot_pressure}
\title{Plot pressure}
\usage{
plot_pressure(
  pressure_data,
  variable = "max",
  smooth = FALSE,
  frame,
  plot_COP = FALSE,
  plot_outline = FALSE,
  plot = TRUE
)
}
\arguments{
\item{pressure_data}{List. Includes a 3D array covering each timepoint of the
measurement. z dimension represents time}

\item{variable}{String. "max" = footprint of maximum sensors. "mean" = average
value of sensors over time (usually for static analyses). "frame" = an
individual frame}

\item{smooth}{Logical. If TRUE, plot will interpolate between sensors to
increase data density}

\item{frame}{Integer.}

\item{plot_COP}{Logical. If TRUE, overlay COP data on plot. Default = FALSE}

\item{plot_outline}{Logical. If TRUE, overlay convex hull outline on plot}

\item{plot}{Logical. If TRUE, plot will be displayed}
}
\value{
ggplot plot object
}
\description{
Produce plot of pressure data
}
\examples{
plot_pressure(pressure_data, variable = "max", plot_COP = TRUE)
}
\author{
Scott Telfer \email{scott.telfer@gmail.com}
}
