% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pressuRe-package.R
\docType{package}
\name{pressuRe-package}
\alias{pressuRe}
\alias{pressuRe-package}
\title{pressuRe: Imports, processes, and visualizes biomechanical pressure data}
\description{
Allows biomechanical pressure data from a range of systems to be imported and processed in a repeatable manner. Also includes functions for visualizing and animating data.
}
\author{
\strong{Maintainer}: Scott Telfer \email{scott.telfer@gmail.com} (\href{https://orcid.org/0000-0002-0104-4027}{ORCID})

}
\keyword{internal}
