% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pressuRe_functions.R
\name{pressure_curve}
\alias{pressure_curve}
\title{Pressure curve}
\usage{
pressure_curve(pressure_data, variable = "peak", plot = FALSE)
}
\arguments{
\item{pressure_data}{List. First item should be a 3D array covering each
timepoint of the measurement. z dimension represents time}

\item{variable}{String. Whether "peak", "mean", or "pti" should be used}

\item{plot}{Logical. If TRUE also plots data as line curve}
}
\value{
Numeric vector containing force values
}
\description{
Generate pressure curve with option to plot
}
\examples{
pressure_data <- load_emed("inst/extdata/emed_test.lst")
pressure_curve(pressure_data, variable = "peak", plot = TRUE)
}
\author{
Scott Telfer \email{scott.telfer@gmail.com}
}
