% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pressuRe_functions.R
\name{pressure_outline}
\alias{pressure_outline}
\title{Get outline of pressure region
Determine outline (convex hull) of pressure measurement}
\usage{
pressure_outline(pressure_data, frame)
}
\arguments{
\item{pressure_data}{List. Includes a 3D array covering each timepoint of the
measurement. z dimension represents time}

\item{frame}{Integer. Frame number to use}
}
\value{
Polygon. sfg object representing convex hull outline
}
\description{
Get outline of pressure region
Determine outline (convex hull) of pressure measurement
}
\examples{
emed_data <- system.file("extdata", "emed_test.lst", package = "pressuRe")
pressure_data <- load_emed(emed_data)
pressure_outline(pressure_data)
pressure_outline(pressure_data, frame = 50)
}
\author{
Scott Telfer \email{scott.telfer@gmail.com}
}
