% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pressuRe_functions.R
\name{select_steps}
\alias{select_steps}
\title{Select steps}
\usage{
select_steps(
  pressure_data,
  threshold_R = 20,
  threshold_L = 20,
  min_frames = 2,
  steps_Rn = 12,
  steps_Ln = 12
)
}
\arguments{
\item{threshold_R}{Numeric. Threshold force to define start and end of step}

\item{threshold_L}{Numeric. Threshold force to define start and end of step}

\item{min_frames}{Numeric. Minimum number of frames that need to be in step}

\item{steps_Rn}{Numeric. Target number of steps for right foot}

\item{steps_Ln}{Numeric. Target number of steps for left foot}

\item{pressure_data.}{List}
}
\value{
\itemize{
  \item pressure_array. 3D array covering each timepoint of the measurement.
           z dimension represents time
  \item pressure_system. Character defining pressure system
  \item sens_size. Numeric vector with the dimensions of the sensors
  \item time. Numeric value for time between measurements
  \item masks. List
  \item events. List
  }
}
\description{
Select steps, usually from insole data, and format for analysis
}
\examples{
select_steps
}
