% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pressuRe_functions.R
\name{sensor_coords}
\alias{sensor_coords}
\title{Coordinates of active sensors}
\usage{
sensor_coords(pressure_data, active = FALSE)
}
\arguments{
\item{pressure_data}{List. Includes a 3D array covering each timepoint of the
measurement. z dimension represents time}

\item{active.}{Logical. Should the coordinates be limited to active sensors}
}
\value{
Data frame. x and y coordinates of sensors
}
\description{
Coordinates of active sensors
}
\examples{
sensor_coords(pressure_data)
}
\author{
Scott Telfer \email{scott.telfer@gmail.com}
}
