% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pressuRe_functions.R
\name{toe_line}
\alias{toe_line}
\title{Get toe line}
\usage{
toe_line(pressure_data, side)
}
\arguments{
\item{pressure_data}{List. First item should be a 3D array covering each
timepoint of the measurement. z dimension represents time}

\item{side}{String}
}
\description{
Calculates line proximal to toes
}
