% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pressuRe_functions.R
\name{animate_pressure}
\alias{animate_pressure}
\title{Animate pressure}
\usage{
animate_pressure(
  pressure_data,
  plot_colors = "default",
  fps,
  dpi = 96,
  file_name,
  preview = FALSE
)
}
\arguments{
\item{pressure_data}{Array. A 3D array covering each timepoint of the
measurement. z dimension represents time}

\item{plot_colors}{String}

\item{fps}{Numeric. Number of frames per second in animation}

\item{dpi}{Numeric. Resolution of gif}

\item{file_name}{Name (inlcuding path) of export file}

\item{preview}{Logical. Whether to play the animation}
}
\value{
Animation in gif format
}
\description{
Produce animation (gif) of pressure distribution
}
\examples{
pressure_data <- load_emed("inst/extdata/emed_test.lst")
animate_pressure(pressure_data, fps = 10, filename = "testgif3.gif")
}
\author{
Scott Telfer \email{scott.telfer@gmail.com}
}
