% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pressuRe_functions.R
\name{animate_pressure}
\alias{animate_pressure}
\title{Animate pressure}
\usage{
animate_pressure(pressure_data, fps, filename, preview = FALSE)
}
\arguments{
\item{pressure_data}{Array. A 3D array covering each timepoint of the
measurement. z dimension represents time}

\item{fps}{Numeric. Number of frames per second in animation}

\item{preview}{Logical. Whether to play the animation}

\item{file}{Name (inlcuding path) of export file}
}
\value{
Animation in gif format
}
\description{
Produce animation (gif) of pressure distribution
}
\examples{
animate_pressure(pressure_data, fps = 10, "testgif.gif")
}
\author{
Scott Telfer \email{scott.telfer@gmail.com}
}
