% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pressuRe_functions.R
\name{automask}
\alias{automask}
\title{automask footprint}
\usage{
automask(pressure_data, foot_side, sens = 4, plot = FALSE)
}
\arguments{
\item{pressure_data}{List. First item is a 3D array covering each timepoint
of the measurement. z dimension represents time}

\item{foot_side}{Character. "RIGHT", "LEFT", or "auto". Auto uses
auto_detect_side function}

\item{sens}{Numeric. Number of frames per second in animation}

\item{plot}{Logical. Whether to play the animation}
}
\value{
List. Contains polygon with each mask
}
\description{
Automatically create mask of footprint data
}
\examples{
pressure_data <- load_emed("inst/extdata/emed_test.lst")
masks <- automask(pressure_data, foot_side = "auto", sens = 4, plot = TRUE)
}
\author{
Scott Telfer \email{scott.telfer@gmail.com}
}
