% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pressuRe_functions.R
\name{automask}
\alias{automask}
\title{automask footprint}
\usage{
automask(pressure_data, side, sens = 4, plot = FALSE)
}
\arguments{
\item{pressure_data}{List. First item is a 3D array covering each timepoint
of the measurement. z dimension represents time}

\item{side}{Character. "RIGHT" or "LEFT"}

\item{sens}{Numeric. Number of frames per second in animation}

\item{plot}{Logical. Whether to play the animation}
}
\value{
List. Contains polygon with each mask
}
\description{
Automatically create mask of footprint data
}
\examples{
automask(pressure_data, sens = 4, plot = TRUE)
}
\author{
Scott Telfer \email{scott.telfer@gmail.com}
}
