% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pressuRe_functions.R
\name{create_mask}
\alias{create_mask}
\title{Create mask}
\usage{
create_mask(pressure_data, image = c())
}
\arguments{
\item{pressure_data}{List. First item is a 3D array covering each timepoint
of the measurement.}

\item{image_value}{String. "max" = footprint of maximum sensors. "mean"
average value of sensors over time (usually for static analyses)}
}
\value{
Array. A 3D array covering each timepoint of the measurement for the
  selected region. z dimension represents time
}
\description{
Manually create mask region
}
\author{
Scott Telfer \email{scott.telfer@gmail.com}
}
