% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pressuRe_functions.R
\name{dpli}
\alias{dpli}
\title{Dynamic Plantar Loading Index}
\usage{
dpli(pressure_data)
}
\arguments{
\item{pressure_data}{List. First item is a 3D array covering each timepoint
of the measurement.}
}
\description{
Determine Dynamic plantar loading index (DPLI) for
footprint pressure data
}
\author{
Scott Telfer \email{scott.telfer@gmail.com}
}
