% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pressuRe_functions.R
\name{edit_mask}
\alias{edit_mask}
\title{Edit mask}
\usage{
edit_mask(pressure_data)
}
\arguments{
\item{pressure_data}{List. First item is a 3D array covering each timepoint
of the measurement.}
}
\value{
List.
}
\description{
Manually edit mask
}
\examples{
edit_mask(pressure_data)
}
\author{
Scott Telfer \email{scott.telfer@gmail.com}
}
