% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pressuRe_functions.R
\name{edit_mask}
\alias{edit_mask}
\title{Edit mask}
\usage{
edit_mask(pressure_data, n_edit)
}
\arguments{
\item{pressure_data}{List. First item is a 3D array covering each timepoint
of the measurement.}

\item{n_edit}{Numeric. Number of vertices to edit}
}
\value{
List.
}
\description{
Manually edit mask
}
\examples{
\dontrun{
emed_data <- system.file("extdata", "emed_test.lst", package = "pressuRe")
pressure_data <- load_emed(emed_data)
pressure_data <- automask(pressure_data, foot_side = "auto", plot = TRUE)
edit_mask(pressure_data)
}
}
\author{
Scott Telfer \email{scott.telfer@gmail.com}
}
