% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pressuRe_functions.R
\name{force_curve}
\alias{force_curve}
\title{Force curve}
\usage{
force_curve(pressure_data, side, plot = FALSE)
}
\arguments{
\item{pressure_data}{List. A 3D array covering each timepoint of the
measurement. z dimension represents time}

\item{side}{For insole data only}

\item{plot}{Logical. If TRUE also plots data as line curve}
}
\value{
Numeric vector containing force values
}
\description{
Generates force vector with option to plot
}
\examples{
pressure_data <- load_emed("inst/extdata/emed_test.lst")
force_curve(pressure_data, side = "left", plot = TRUE)
}
\author{
Scott Telfer \email{scott.telfer@gmail.com}
}
