% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pressuRe_functions.R
\name{load_emed}
\alias{load_emed}
\title{Load emed data}
\usage{
load_emed(pressure_filepath, rem_zeros = TRUE)
}
\arguments{
\item{pressure_filepath}{String. Filepath pointing to emed pressure file}

\item{rem_zeros}{Logical. If TRUE, remove columns and rows from edges which
contain only zeros}
}
\value{
A list with information about the pressure data.
\itemize{
  \item pressure_array. A 3D array covering each timepoint of the measurement.
           z dimension represents time
  \item pressure_system. String with the name of the system the data was
           collected on
  \item sens_size. Numeric vector with dimensions of the sensors (m)
  \item time. Numeric value for time between measurements
 }
}
\description{
Imports and formats .lst files collected on emed system and
   exported from Novel software
}
\examples{
pressure_data = load_emed("inst/extdata/emed_test.lst")
}
\author{
Scott Telfer \email{scott.telfer@gmail.com}
}
