% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pressuRe_functions.R
\name{load_fscan}
\alias{load_fscan}
\title{Load F-scan data}
\usage{
load_fscan(pressure_filepath)
}
\arguments{
\item{pressure_filepath}{String. Filepath pointing to emed pressure file}
}
\value{
A list with information about the pressure data.
\itemize{
  \item pressure_array. 3D array covering each timepoint of the measurement.
           z dimension represents time
  \item pressure_system. Character defining pressure system
  \item sens_size. Numeric vector with the dimensions of the sensors
  \item time. Numeric value for time between measurements
  \item masks. List
  \item events. List
 }
 @examples
 pressure_data <- load_fscan("inst/extdata/)
 @importFrom
 @export
}
\description{
Imports and formats .asc files collected on f-scan system and
   exported from Tekscan software
}
\author{
Scott Telfer \email{scott.telfer@gmail.com}
}
