% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pressuRe_functions.R
\name{load_pedar}
\alias{load_pedar}
\title{Load pedar data}
\usage{
load_pedar(pressure_filepath)
}
\arguments{
\item{pressure_filepath}{String. Filepath pointing to emed pressure file}
}
\value{
A list with information about the pressure data.
\itemize{
  \item pressure_array. A 3D array covering each timepoint of the measurement.
           z dimension represents time
  \item pressure_system. String with the name of the system the data was
           collected on
  \item sens_size. Numeric vector with dimensions of the sensors (m), or
           string with name of insole size
  \item time. Numeric value for time between measurements
 }
}
\description{
Imports and formats .asc files collected on pedar system and
   exported from Novel software
}
\examples{
pressure_data = load_pedar("inst/extdata/pedar_example.asc")
}
\author{
Scott Telfer \email{scott.telfer@gmail.com}
}
