% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pressuRe-package.R
\docType{package}
\name{pressuRe-package}
\alias{pressuRe}
\alias{pressuRe-package}
\title{pressuRe: Imports, Processes, and Visualizes Biomechanical Pressure Data}
\description{
Allows biomechanical pressure data from a range of systems to be imported and processed in a reproducible manner. Automatic and manual tools are included to let the user define regions (masks) to be analyzed. Also includes functions for visualizing and animating pressure data. Example methods are described in Shi et al., (2022) \doi{10.1038/s41598-022-19814-0}, Lee et al., (2014) \doi{10.1186/1757-1146-7-18}, van der Zward et al., (2014) \doi{10.1186/1757-1146-7-20}, and Najafi et al., (2010) \doi{10.1016/j.gaitpost.2009.09.003}.
}
\author{
\strong{Maintainer}: Scott Telfer \email{scott.telfer@gmail.com} (\href{https://orcid.org/0000-0002-0104-4027}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Ellen Li \email{ellenlii07@gmail.com} (\href{https://orcid.org/0000-0001-5545-7364}{ORCID})
}

}
\keyword{internal}
