% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pressuRe_functions.R
\name{pressure_outline}
\alias{pressure_outline}
\title{Determine outline (convex hull) of pressure measurement}
\usage{
pressure_outline(pressure_data)
}
\arguments{
\item{pressure_data}{List. Includes a 3D array covering each timepoint of the
measurement. z dimension represents time}
}
\value{
Data frame. x and y coordinators of convex hull outline
}
\description{
Determine outline (convex hull) of pressure measurement
}
\author{
Scott Telfer \email{scott.telfer@gmail.com}
}
